Show local verification symbol on Contract tab
